Use double for volume

+++ src/scratch/target.cpp
@@ -338,13 +338,13 @@ void Target::setLayerOrder(int newLayerOrder)
 }
 
 /*! Returns the volume. */
-int Target::volume() const
+double Target::volume() const
 {
     return impl->volume;
 }
 
 /*! Sets the volume. */
-void Target::setVolume(int newVolume)
+void Target::setVolume(double newVolume)
 {
     impl->volume = newVolume;
 }
